% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HFfunctions.R
\name{init_sq_mat}
\alias{init_sq_mat}
\title{init_sq_mat}
\usage{
init_sq_mat(m)
}
\arguments{
\item{m}{basis size}
}
\value{
square matrix
}
\description{
init_sq_mat
}
\examples{
Usage example: R, Rold, C, Cbar, and HF are all square matrices with dim=(basis size, basis size)
}
